requirements.txt
# ============================================================
# Medical Chatbot - Python Dependencies
# Thesis: Design and Implementation of a Task-Oriented 
#         Medical Chatbot using Llama 3 8B
# ============================================================

# ============================================================
# Core Web Framework
# ============================================================
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Session==0.5.0

# ============================================================
# Database
# ============================================================
PyMySQL==1.1.0
SQLAlchemy==2.0.23
alembic==1.12.1

# ============================================================
# Environment and Configuration
# ============================================================
python-dotenv==1.0.0

# ============================================================
# HTTP Client for API Calls
# ============================================================
requests==2.31.0
urllib3==2.0.7

# ============================================================
# Security and Input Sanitization
# ============================================================
bleach==6.1.0
markdown==3.5.1
email-validator==2.1.0

# ============================================================
# AI/ML - Llama 3 8B (Optional - comment out if not using local model)
# ============================================================
# transformers==4.36.0
# torch==2.1.0
# accelerate==0.25.0
# sentencepiece==0.1.99
# protobuf==4.25.0

# ============================================================
# Testing
# ============================================================
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.2.0

# ============================================================
# Development Tools
# ============================================================
black==23.12.1
flake8==7.0.0
mypy==1.7.1

# ============================================================
# Production Server
# ============================================================
gunicorn==21.2.0

# ============================================================
# Logging and Monitoring
# ============================================================
python-json-logger==2.0.7

# ============================================================
# Data Processing
# ============================================================
python-dateutil==2.8.2